refactor: calculation page
