+++ src/components/PersonList/PersonList.module.css
@@ -1,0 +1,22 @@
+.container {
+  display: flex;
+  flex-direction: row;
+}
+
+.wrap {
+  width: 64px;
+  height: 64px;
+  position: relative;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.wrap > img {
+  width: 80%;
+  height: 80%;
+  border-radius: 50%;
+  border: 1px solid black;
+  padding: 5%;
+}
