+++ src/components/TotalPrice/TotalPrice.module.css
@@ -1,0 +1,11 @@
+.container {
+  font-size: 3rem;
+  margin-top: 30px;
+}
+
+.wrap {
+  display: flex;
+  justify-content: space-between;
+  font-size: 2rem;
+  margin: 30px 20px 0 20px;
+}
