+++ src/pages/Calculation/Calculation.module.css
@@ -1,0 +1,4 @@
+.container {
+  height: calc(100vh - 250px);
+  margin: 25px;
+}
